#130, #148: Removed ConstantSlice.
