+++ core/src/main/java/io/parsingdata/metal/Shorthand.java
@@ -16,7 +16,7 @@
 
 package io.parsingdata.metal;
 
-import static io.parsingdata.metal.data.ConstantSlice.create;
+import static io.parsingdata.metal.expression.value.ConstantFactory.makeConstantSlice;
 import static io.parsingdata.metal.token.Token.NO_NAME;
 
 import java.util.function.BinaryOperator;
@@ -146,7 +146,7 @@ public final class Shorthand {
     public static ValueExpression con(final String value) { return con(value, new Encoding()); }
     public static ValueExpression con(final String value, final Encoding encoding) { return con(ConstantFactory.createFromString(value, encoding)); }
     public static ValueExpression con(final Value value) { return new Const(value); }
-    public static ValueExpression con(final Encoding encoding, final int... values) { return new Const(new Value(create(toByteArray(values)), encoding)); }
+    public static ValueExpression con(final Encoding encoding, final int... values) { return new Const(new Value(makeConstantSlice(toByteArray(values)), encoding)); }
     public static ValueExpression con(final int... values) { return con(new Encoding(), values); }
     public static ValueExpression con(final byte[] value) { return con(value, new Encoding()); }
     public static ValueExpression con(final byte[] value, final Encoding encoding) { return con(ConstantFactory.createFromBytes(value, encoding)); }
