+++ core/src/main/java/io/parsingdata/metal/expression/value/ConstantFactory.java
@@ -16,11 +16,11 @@
 
 package io.parsingdata.metal.expression.value;
 
-import static io.parsingdata.metal.data.ConstantSlice.create;
-
 import java.math.BigInteger;
 import java.util.BitSet;
 
+import io.parsingdata.metal.data.ConstantSource;
+import io.parsingdata.metal.data.Slice;
 import io.parsingdata.metal.encoding.ByteOrder;
 import io.parsingdata.metal.encoding.Encoding;
 import io.parsingdata.metal.encoding.Sign;
@@ -29,8 +29,12 @@ public final class ConstantFactory {
 
     private ConstantFactory() {}
 
+    public static Slice makeConstantSlice(final byte[] data) {
+        return new Slice(new ConstantSource(data), 0, data);
+    }
+
     public static Value createFromBytes(final byte[] value, final Encoding encoding) {
-        return new Value(create(value), encoding);
+        return new Value(makeConstantSlice(value), encoding);
     }
 
     public static Value createFromNumeric(final BigInteger value, final Encoding encoding) {
@@ -42,14 +46,14 @@ public final class ConstantFactory {
     }
 
     public static Value createFromString(final String value, final Encoding encoding) {
-        return new Value(create(value.getBytes(encoding.charset)), encoding);
+        return new Value(makeConstantSlice(value.getBytes(encoding.charset)), encoding);
     }
 
     public static Value createFromBitSet(final BitSet value, final int minSize, final Encoding encoding) {
         final byte[] bytes = ByteOrder.LITTLE_ENDIAN.apply(value.toByteArray());
         final byte[] outBytes = new byte[Math.max(minSize, bytes.length)];
         System.arraycopy(bytes, 0, outBytes, outBytes.length - bytes.length, bytes.length);
-        return new Value(create(outBytes), setToBigEndian(encoding));
+        return new Value(makeConstantSlice(outBytes), setToBigEndian(encoding));
     }
 
     private static Encoding setToBigEndian(final Encoding encoding) {
