+++ core/src/test/java/io/parsingdata/metal/ToStringTest.java
@@ -55,9 +55,9 @@ import static io.parsingdata.metal.Shorthand.sub;
 import static io.parsingdata.metal.Shorthand.tie;
 import static io.parsingdata.metal.Shorthand.token;
 import static io.parsingdata.metal.Shorthand.whl;
-import static io.parsingdata.metal.data.ConstantSlice.create;
 import static io.parsingdata.metal.data.ParseGraph.NONE;
 import static io.parsingdata.metal.data.selection.ByName.getValue;
+import static io.parsingdata.metal.expression.value.ConstantFactory.makeConstantSlice;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
 import static io.parsingdata.metal.util.TokenDefinitions.any;
@@ -157,12 +157,12 @@ public class ToStringTest {
         assertEquals(envString, environment.toString());
         final ParseResult result = new ParseResult(true, environment);
         assertEquals("ParseResult(true, " + environment + ")", result.toString());
-        final ParseValue pv1 = new ParseValue("name", NONE, create(new byte[]{1, 2}), enc());
+        final ParseValue pv1 = new ParseValue("name", NONE, makeConstantSlice(new byte[]{1, 2}), enc());
         final String pv1String = "name(0x0102)";
         final Optional<Value> ov1 = Optional.of(pv1);
-        final Optional<Value> ov2 = Optional.of(new Value(create(new byte[]{3}), enc()));
+        final Optional<Value> ov2 = Optional.of(new Value(makeConstantSlice(new byte[]{3}), enc()));
         assertEquals(">Optional[0x03]>Optional[" + pv1String + "]", ImmutableList.create(ov1).add(ov2).toString());
-        final ParseValue pv2 = new ParseValue("two", NONE, create(new byte[]{3, 4}), enc());
+        final ParseValue pv2 = new ParseValue("two", NONE, makeConstantSlice(new byte[]{3, 4}), enc());
         final String pv2String = "two(0x0304)";
         assertEquals(">" + pv2String + ">" + pv1String, ImmutableList.create(pv1).add(pv2).toString());
         assertEquals(">" + pv2String + ">" + pv1String, ImmutableList.create(pv1).add(pv2).toString());
@@ -170,10 +170,10 @@ public class ToStringTest {
 
     @Test
     public void slice() throws IOException {
-        final ParseValue pv1 = new ParseValue("name", NONE, create(new byte[]{1, 2}), enc());
+        final ParseValue pv1 = new ParseValue("name", NONE, makeConstantSlice(new byte[]{1, 2}), enc());
         assertEquals("0102@0:2", pv1.slice.toString());
         final Environment oneValueEnvironment = stream().add(pv1);
-        final Environment twoValueEnvironment = oneValueEnvironment.add(new ParseValue("name2", NONE, new DataExpressionSource(ref("name"), 0, oneValueEnvironment, enc()).slice(0, 2), enc()));
+        final Environment twoValueEnvironment = oneValueEnvironment.add(new ParseValue("name2", NONE, new DataExpressionSource(ref("name"), 0, oneValueEnvironment.order, enc()).slice(0, 2), enc()));
         final String dataExpressionSliceString = getValue(twoValueEnvironment.order, "name2").slice.toString();
         assertTrue(dataExpressionSliceString.startsWith("NameRef(name)[0]("));
         assertTrue(dataExpressionSliceString.endsWith(")@0:2"));
