+++ core/src/test/java/io/parsingdata/metal/data/ParseValueTest.java
@@ -23,7 +23,7 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 import static io.parsingdata.metal.Shorthand.def;
-import static io.parsingdata.metal.data.ConstantSlice.create;
+import static io.parsingdata.metal.expression.value.ConstantFactory.makeConstantSlice;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static junit.framework.TestCase.assertFalse;
 
@@ -45,7 +45,7 @@ public class ParseValueTest {
     @Before
     public void setUp() {
         definition = def("value", 1);
-        value = new ParseValue("value", definition, create(new byte[] { 1 }), enc());
+        value = new ParseValue("value", definition, makeConstantSlice(new byte[] { 1 }), enc());
     }
 
     @Test
