+++ formats/src/main/java/io/parsingdata/metal/format/Callback.java
@@ -16,7 +16,7 @@
 
 package io.parsingdata.metal.format;
 
-import static io.parsingdata.metal.data.ConstantSlice.create;
+import static io.parsingdata.metal.expression.value.ConstantFactory.makeConstantSlice;
 
 import java.util.Optional;
 import java.util.zip.CRC32;
@@ -38,10 +38,11 @@ public final class Callback {
                 final CRC32 crc = new CRC32();
                 crc.update(value.getValue());
                 final long crcValue = crc.getValue();
-                return Optional.of(new Value(create(encoding.byteOrder.apply(new byte[] { (byte)((crcValue & 0xff000000) >> 24),
-                                                                                          (byte)((crcValue & 0xff0000) >> 16),
-                                                                                          (byte)((crcValue & 0xff00) >> 8),
-                                                                                          (byte) (crcValue & 0xff) })), encoding));
+                return Optional.of(new Value(makeConstantSlice(encoding.byteOrder.apply(new byte[] {
+                    (byte)((crcValue & 0xff000000) >> 24),
+                    (byte)((crcValue & 0xff0000) >> 16),
+                    (byte)((crcValue & 0xff00) >> 8),
+                    (byte) (crcValue & 0xff) })), encoding));
             }
         };
     }
